feat(views): implement vendor list and detail
