+++ src/views/Vendors/Vendors.tsx
@@ -1,46 +1,34 @@
 /**
  * @module sbom-harbor-ui/views/Vendors/Vendors
  */
-import { Suspense } from 'react'
+import { Suspense, useCallback } from 'react'
 import { useQuery } from 'react-query'
-import Box from '@mui/material/Box'
-import CircularProgress from '@mui/material/CircularProgress'
 import Typography from '@mui/material/Typography'
-import List from '@/components/crud/List'
 import LinearIndeterminate from '@/components/mui/LinearLoadingBar'
+import List from '@/components/crud/List'
 import { vendorsQuery } from '@/router/vendorsLoader'
+import schema from './schema'
 
-const schema = [
-  {
-    name: 'name',
-    label: 'Name',
-    type: 'text',
-  },
-  {
-    name: 'address',
-    label: 'Address',
-    type: 'text',
-  },
-]
-
+/**
+ * React Suspense-aware Component for rendering all vendors in a list view.
+ * @return {JSX.Element} the vendors list view
+ */
 const VendorsContainer = (): JSX.Element => {
   const { data: vendors = [], isLoading } = useQuery(vendorsQuery())
 
-  const deleteVendor = (id: string) => {
-    // TODO: implement delete vendor functionality
-    console.debug('deleteVendor: ', id)
-  }
+  const handleDelete = useCallback((id: string) => {
+    console.debug('handleDelete: ', id)
+  }, [])
 
   return (
-    <Box>
+    <>
       <Typography variant="h4" sx={{ mb: 2 }}>
-        Vendors
+        Vendors List
       </Typography>
-      {isLoading && <CircularProgress />}
       <Suspense fallback={<LinearIndeterminate />}>
-        <List items={vendors || []} schema={schema} deleteItem={deleteVendor} />
+        <List items={vendors || []} schema={schema} deleteItem={handleDelete} />
       </Suspense>
-    </Box>
+    </>
   )
 }
